@@ -7,9 +7,10 @@
 //  <last-date>2018-06-27 4:44</last-date>
 // -----------------------------------------------------------------------
 
+using System;
 using System.ComponentModel;
 
-using OSharp.Identity;
+using OSharp.Identity.Entities;
 
 
 namespace Liuliu.Demo.Identity.Entities
@@ -18,7 +19,7 @@ namespace Liuliu.Demo.Identity.Entities
     /// 实体类：用户角色信息
     /// </summary>
     [Description("用户角色信息")]
-    public class UserRole : UserRoleBase<int, int>
+    public class UserRole : UserRoleBase<Guid, int, int>
     {
         /// <summary>
         /// 获取或设置 关联用户信息
