@@ -1,10 +1,10 @@
 ﻿// -----------------------------------------------------------------------
 //  <copyright file="UserStore.cs" company="OSharp开源团队">
-//      Copyright (c) 2014-2018 OSharp. All rights reserved.
+//      Copyright (c) 2014-2020 OSharp. All rights reserved.
 //  </copyright>
 //  <site>http://www.osharp.org</site>
 //  <last-editor>郭明锋</last-editor>
-//  <last-date>2018-06-27 4:44</last-date>
+//  <last-date>2020-01-31 19:34</last-date>
 // -----------------------------------------------------------------------
 
 using System;
@@ -21,10 +21,10 @@ namespace Liuliu.Demo.Identity
     /// <summary>
     /// 用户仓储
     /// </summary>
-    public class UserStore : UserStoreBase<User, int, UserClaim, UserLogin, UserToken, Role, int, UserRole>
+    public class UserStore : UserStoreBase<User, int, UserClaim, int, UserLogin, Guid, UserToken, Guid, Role, int, UserRole, Guid>
     {
         /// <summary>
-        /// 初始化一个<see cref="UserStoreBase{TUser, TUserKey, TUserClaim, TUserLogin, TUserToken, TRole, TRoleKey, TUserRole}"/>类型的新实例
+        /// 初始化一个<see cref="UserStoreBase{TUser, TUserKey, TUserClaim, TUserClaimKey, TUserLogin, TUserLoginKey, TUserToken, TUserTokenKey, TRole, TRoleKey, TUserRole, TUserRoleKey}"/>类型的新实例
         /// </summary>
         /// <param name="userRepository">用户仓储</param>
         /// <param name="userLoginRepository">用户登录仓储</param>
