@@ -25,7 +25,7 @@ namespace Liuliu.Demo.EntityConfiguration.Systems
         {
             builder.HasData(
                 new KeyValue() { Id = Guid.Parse("534d7813-0eea-44cc-b88e-a9cb010c6981"), Key = SystemSettingKeys.SiteName, Value = "OSHARP" },
-                new KeyValue() { Id = Guid.Parse("977e4bba-97b2-4759-a768-a9cb010c698c"), Key = SystemSettingKeys.SiteDescription, Value = "Osharp with .NetStandard2.0 & Angular6" }
+                new KeyValue() { Id = Guid.Parse("977e4bba-97b2-4759-a768-a9cb010c698c"), Key = SystemSettingKeys.SiteDescription, Value = "Osharp with AspNetCore & Angular" }
             );
         }
     }
