@@ -9,7 +9,7 @@
 
 using System.ComponentModel;
 
-using Liuliu.Demo.Security;
+using Liuliu.Demo.Authorization;
 
 using Microsoft.AspNetCore.Mvc;
 
@@ -19,13 +19,6 @@ namespace Liuliu.Demo.Web.Areas.Admin.Controllers
     [Description("管理-主页")]
     public class HomeController : AdminApiController
     {
-        private readonly SecurityManager _securityManager;
-
-        public HomeController(SecurityManager securityManager)
-        {
-            _securityManager = securityManager;
-        }
-
         /// <summary>
         /// 获取后台管理主菜单
         /// </summary>
