@@ -21,11 +21,11 @@ using Microsoft.AspNetCore.Mvc;
 
 using OSharp.AspNetCore.Mvc.Filters;
 using OSharp.AspNetCore.UI;
-using OSharp.Core.Modules;
+using OSharp.Authorization;
+using OSharp.Authorization.Modules;
 using OSharp.Data;
 using OSharp.Entity;
 using OSharp.Filter;
-using OSharp.Security;
 
 
 namespace Liuliu.Demo.Web.Areas.Admin.Controllers
