@@ -27,8 +27,8 @@ using OSharp.AspNetCore;
 using OSharp.AspNetCore.Mvc;
 using OSharp.AspNetCore.Mvc.Filters;
 using OSharp.AspNetCore.UI;
-using OSharp.Core;
-using OSharp.Core.Modules;
+using OSharp.Authorization;
+using OSharp.Authorization.Modules;
 using OSharp.Data;
 using OSharp.Entity;
 using OSharp.Extensions;
@@ -39,7 +39,6 @@ using OSharp.Identity.OAuth2;
 using OSharp.Json;
 using OSharp.Mapping;
 using OSharp.Net;
-using OSharp.Security.Claims;
 
 
 namespace Liuliu.Demo.Web.Controllers
@@ -73,11 +72,7 @@ namespace Liuliu.Demo.Web.Controllers
         [Description("用户名是否存在")]
         public bool CheckUserNameExists(string userName)
         {
-#if NETCOREAPP3_0
             bool exists = _userManager.Users.Any(m => m.NormalizedUserName == _userManager.NormalizeName(userName));
-#else
-            bool exists = _userManager.Users.Any(m => m.NormalizedUserName == _userManager.NormalizeKey(userName));
-#endif
             return exists;
         }
 
@@ -90,11 +85,7 @@ namespace Liuliu.Demo.Web.Controllers
         [Description("用户Email是否存在")]
         public bool CheckEmailExists(string email)
         {
-#if NETCOREAPP3_0
             bool exists = _userManager.Users.Any(m => m.NormalizeEmail == _userManager.NormalizeEmail(email));
-#else
-            bool exists = _userManager.Users.Any(m => m.NormalizeEmail == _userManager.NormalizeKey(email));
-#endif
             return exists;
         }
 
@@ -206,40 +197,6 @@ namespace Liuliu.Demo.Web.Controllers
         }
 
         /// <summary>
-        /// Jwt登录
-        /// </summary>
-        /// <param name="dto">登录信息</param>
-        /// <returns>JSON操作结果</returns>
-        [HttpPost]
-        [ModuleInfo]
-        [Description("JWT登录")]
-        public async Task<AjaxResult> Jwtoken(LoginDto dto)
-        {
-            Check.NotNull(dto, nameof(dto));
-
-            if (!ModelState.IsValid)
-            {
-                return new AjaxResult("提交信息验证失败", AjaxResultType.Error);
-            }
-
-            dto.Ip = HttpContext.GetClientIp();
-            dto.UserAgent = Request.Headers["User-Agent"].FirstOrDefault();
-
-            OperationResult<User> result = await _identityContract.Login(dto);
-            IUnitOfWork unitOfWork = HttpContext.RequestServices.GetUnitOfWork<User, int>();
-            unitOfWork.Commit();
-
-            if (!result.Succeeded)
-            {
-                return result.ToAjaxResult();
-            }
-
-            User user = result.Data;
-            JsonWebToken token = await CreateJwtToken(user);
-            return new AjaxResult("登录成功", AjaxResultType.Success, token);
-        }
-
-        /// <summary>
         /// 获取身份认证Token
         /// </summary>
         /// <param name="dto">TokenDto</param>
@@ -249,7 +206,8 @@ namespace Liuliu.Demo.Web.Controllers
         [Description("JwtToken")]
         public async Task<AjaxResult> Token(TokenDto dto)
         {
-            if (dto.GrantType == GrantType.Password)
+            string grantType = dto.GrantType?.UpperToLowerAndSplit("_");
+            if (grantType == GrantType.Password)
             {
                 Check.NotNull(dto.Account, nameof(dto.Account));
                 Check.NotNull(dto.Password, nameof(dto.Password));
@@ -271,11 +229,11 @@ namespace Liuliu.Demo.Web.Controllers
                 }
 
                 User user = result.Data;
-                JsonWebToken token = await CreateJwtToken(user);
+                JsonWebToken token = await CreateJwtToken(user, dto.ClientType);
                 return new AjaxResult("登录成功", AjaxResultType.Success, token);
             }
 
-            if (dto.GrantType == GrantType.RefreshToken)
+            if (grantType == GrantType.RefreshToken)
             {
                 Check.NotNull(dto.RefreshToken, nameof(dto.RefreshToken));
                 JsonWebToken token = await CreateJwtToken(dto.RefreshToken);
@@ -409,11 +367,11 @@ namespace Liuliu.Demo.Web.Controllers
             return new AjaxResult("登录成功", AjaxResultType.Success, token);
         }
 
-        private async Task<JsonWebToken> CreateJwtToken(User user)
+        private async Task<JsonWebToken> CreateJwtToken(User user, RequestClientType clientType = RequestClientType.Browser)
         {
             IServiceProvider provider = HttpContext.RequestServices;
             IJwtBearerService jwtBearerService = provider.GetService<IJwtBearerService>();
-            JsonWebToken token = await jwtBearerService.CreateToken(user.Id.ToString(), user.UserName);
+            JsonWebToken token = await jwtBearerService.CreateToken(user.Id.ToString(), user.UserName, clientType);
 
             return token;
         }
