@@ -13,6 +13,7 @@ using Microsoft.AspNetCore.Routing;
 using Microsoft.Extensions.DependencyInjection;
 
 using OSharp.AspNetCore;
+using OSharp.AspNetCore.Routing;
 using OSharp.Core.Packs;
 using OSharp.Exceptions;
 
@@ -54,12 +55,10 @@ namespace Microsoft.AspNetCore.Builder
             });
         }
 
-#if NETCOREAPP3_0
-
         /// <summary>
         /// 添加Endpoint并Area路由支持
         /// </summary>
-        public static IEndpointRouteBuilder MvcEndpointsWithAreaRoute(this IEndpointRouteBuilder endpoints, bool area = true)
+        public static IEndpointRouteBuilder MapControllersWithAreaRoute(this IEndpointRouteBuilder endpoints, bool area = true)
         {
             if (area)
             {
@@ -69,7 +68,5 @@ namespace Microsoft.AspNetCore.Builder
             endpoints.MapDefaultControllerRoute();
             return endpoints;
         }
-
-#endif
     }
 }
