@@ -14,8 +14,8 @@ using System.Reflection;
 
 using Microsoft.AspNetCore.Mvc;
 
-using OSharp.Core.Functions;
-using OSharp.Core.Modules;
+using OSharp.Authorization.Functions;
+using OSharp.Authorization.Modules;
 using OSharp.Exceptions;
 using OSharp.Reflection;
 
