@@ -9,10 +9,11 @@
 
 using System;
 
-using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.SignalR;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Hosting;
 
 using Newtonsoft.Json.Serialization;
 
@@ -48,11 +49,32 @@ namespace OSharp.AspNetCore.SignalR
             services.TryAddSingleton<IUserIdProvider, UserNameUserIdProvider>();
             services.TryAddSingleton<IConnectionUserCache, ConnectionUserCache>();
 
-            ISignalRServerBuilder builder = services.AddSignalR();
+            Action<HubOptions> hubOptions = GetHubOptionsAction(services);
+            ISignalRServerBuilder builder = hubOptions == null
+                ? services.AddSignalR()
+                : services.AddSignalR(hubOptions);
+
             Action<ISignalRServerBuilder> buildAction = GetSignalRServerBuildAction(services);
             buildAction?.Invoke(builder);
 
             return services;
+        }
+
+        /// <summary>
+        /// 重写以获取HubOptions创建委托
+        /// </summary>
+        /// <param name="services">依赖注入服务容器</param>
+        /// <returns></returns>
+        protected virtual Action<HubOptions> GetHubOptionsAction(IServiceCollection services)
+        {
+            return config =>
+            {
+                IWebHostEnvironment environment = services.GetWebHostEnvironment();
+                if (environment.IsDevelopment())
+                {
+                    config.EnableDetailedErrors = true;
+                }
+            };
         }
 
         /// <summary>
@@ -62,32 +84,8 @@ namespace OSharp.AspNetCore.SignalR
         /// <returns></returns>
         protected virtual Action<ISignalRServerBuilder> GetSignalRServerBuildAction(IServiceCollection services)
         {
-#if NETCOREAPP3_0
             return builder => builder.AddNewtonsoftJsonProtocol(options =>
                 options.PayloadSerializerSettings.ContractResolver = new DefaultContractResolver());
-#else
-            return builder => builder.AddJsonProtocol(config =>
-                config.PayloadSerializerSettings.ContractResolver = new DefaultContractResolver());
-#endif
         }
-
-#if NETSTANDARD
-        /// <summary>
-        /// 应用AspNetCore的服务业务
-        /// </summary>
-        /// <param name="app">Asp应用程序构建器</param>
-        public override void UsePack(IApplicationBuilder app)
-        {
-            Action<HubRouteBuilder> hubRouteBuildAction = GetHubRouteBuildAction(app.ApplicationServices);
-            app.UseSignalR(hubRouteBuildAction);
-        }
-
-        /// <summary>
-        /// 重写以获取Hub路由创建委托
-        /// </summary>
-        /// <param name="serviceProvider">服务提供者</param>
-        /// <returns></returns>
-        protected abstract Action<HubRouteBuilder> GetHubRouteBuildAction(IServiceProvider serviceProvider);
-#endif
     }
 }
