@@ -17,6 +17,7 @@ using Microsoft.EntityFrameworkCore;
 
 using OSharp.Exceptions;
 
+using Z.EntityFramework.Extensions;
 using Z.EntityFramework.Plus;
 
 
@@ -116,11 +117,8 @@ namespace OSharp.Entity
             {
                 throw new OsharpException($"参数dbContext类型为“{dbContext.GetType()}”，不能转换为 DbContext");
             }
-#if NETSTANDARD2_1
+
             return context.Set<TEntity>().FromSqlRaw(sql, parameters);
-#else
-            return context.Set<TEntity>().FromSql(new RawSqlString(sql), parameters);
-#endif
         }
     }
 }
