@@ -32,7 +32,7 @@ namespace OSharp.Data
         /// <typeparam name="TException">异常类型</typeparam>
         /// <param name="assertion">要验证的断言。</param>
         /// <param name="message">异常消息。</param>
-        private static void Require<TException>(bool assertion, string message)
+        public static void Required<TException>(bool assertion, string message)
             where TException : Exception
         {
             if (assertion)
@@ -59,7 +59,7 @@ namespace OSharp.Data
             {
                 throw new ArgumentNullException(nameof(assertionFunc));
             }
-            Require<Exception>(assertionFunc(value), message);
+            Required<Exception>(assertionFunc(value), message);
         }
 
         /// <summary>
@@ -76,7 +76,7 @@ namespace OSharp.Data
             {
                 throw new ArgumentNullException("assertionFunc");
             }
-            Require<TException>(assertionFunc(value), message);
+            Required<TException>(assertionFunc(value), message);
         }
 
         /// <summary>
@@ -87,7 +87,7 @@ namespace OSharp.Data
         /// <exception cref="ArgumentNullException"></exception>
         public static void NotNull<T>(T value, string paramName)
         {
-            Require<ArgumentNullException>(value != null, string.Format(Resources.ParameterCheck_NotNull, paramName));
+            Required<ArgumentNullException>(value != null, string.Format(Resources.ParameterCheck_NotNull, paramName));
         }
 
         /// <summary>
@@ -99,7 +99,7 @@ namespace OSharp.Data
         /// <exception cref="ArgumentException"></exception>
         public static void NotNullOrEmpty(string value, string paramName)
         {
-            Require<ArgumentException>(!string.IsNullOrEmpty(value), string.Format(Resources.ParameterCheck_NotNullOrEmpty_String, paramName));
+            Required<ArgumentException>(!string.IsNullOrEmpty(value), string.Format(Resources.ParameterCheck_NotNullOrEmpty_String, paramName));
         }
 
         /// <summary>
@@ -110,7 +110,7 @@ namespace OSharp.Data
         /// <exception cref="ArgumentException"></exception>
         public static void NotEmpty(Guid value, string paramName)
         {
-            Require<ArgumentException>(value != Guid.Empty, string.Format(Resources.ParameterCheck_NotEmpty_Guid, paramName));
+            Required<ArgumentException>(value != Guid.Empty, string.Format(Resources.ParameterCheck_NotEmpty_Guid, paramName));
         }
 
         /// <summary>
@@ -124,7 +124,7 @@ namespace OSharp.Data
         public static void NotNullOrEmpty<T>(IReadOnlyList<T> list, string paramName)
         {
             NotNull(list, paramName);
-            Require<ArgumentException>(list.Any(), string.Format(Resources.ParameterCheck_NotNullOrEmpty_Collection, paramName));
+            Required<ArgumentException>(list.Any(), string.Format(Resources.ParameterCheck_NotNullOrEmpty_Collection, paramName));
         }
 
         /// <summary>
@@ -133,7 +133,7 @@ namespace OSharp.Data
         public static void HasNoNulls<T>(IReadOnlyList<T> list, string paramName)
         {
             NotNull(list, paramName);
-            Require<ArgumentException>(list.All(m => m != null), string.Format(Resources.ParameterCheck_NotContainsNull_Collection, paramName));
+            Required<ArgumentException>(list.All(m => m != null), string.Format(Resources.ParameterCheck_NotContainsNull_Collection, paramName));
         }
 
         /// <summary>
@@ -149,7 +149,7 @@ namespace OSharp.Data
         {
             bool flag = canEqual ? value.CompareTo(target) <= 0 : value.CompareTo(target) < 0;
             string format = canEqual ? Resources.ParameterCheck_NotLessThanOrEqual : Resources.ParameterCheck_NotLessThan;
-            Require<ArgumentOutOfRangeException>(flag, string.Format(format, paramName, target));
+            Required<ArgumentOutOfRangeException>(flag, string.Format(format, paramName, target));
         }
 
         /// <summary>
@@ -165,7 +165,7 @@ namespace OSharp.Data
         {
             bool flag = canEqual ? value.CompareTo(target) >= 0 : value.CompareTo(target) > 0;
             string format = canEqual ? Resources.ParameterCheck_NotGreaterThanOrEqual : Resources.ParameterCheck_NotGreaterThan;
-            Require<ArgumentOutOfRangeException>(flag, string.Format(format, paramName, target));
+            Required<ArgumentOutOfRangeException>(flag, string.Format(format, paramName, target));
         }
 
         /// <summary>
@@ -186,13 +186,13 @@ namespace OSharp.Data
             string message = startEqual
                 ? string.Format(Resources.ParameterCheck_Between, paramName, start, end)
                 : string.Format(Resources.ParameterCheck_BetweenNotEqual, paramName, start, end, start);
-            Require<ArgumentOutOfRangeException>(flag, message);
+            Required<ArgumentOutOfRangeException>(flag, message);
 
             flag = endEqual ? value.CompareTo(end) <= 0 : value.CompareTo(end) < 0;
             message = endEqual
                 ? string.Format(Resources.ParameterCheck_Between, paramName, start, end)
                 : string.Format(Resources.ParameterCheck_BetweenNotEqual, paramName, start, end, end);
-            Require<ArgumentOutOfRangeException>(flag, message);
+            Required<ArgumentOutOfRangeException>(flag, message);
         }
 
         /// <summary>
@@ -205,7 +205,7 @@ namespace OSharp.Data
         public static void DirectoryExists(string directory, string paramName = null)
         {
             NotNull(directory, paramName);
-            Require<DirectoryNotFoundException>(Directory.Exists(directory), string.Format(Resources.ParameterCheck_DirectoryNotExists, directory));
+            Required<DirectoryNotFoundException>(Directory.Exists(directory), string.Format(Resources.ParameterCheck_DirectoryNotExists, directory));
         }
 
         /// <summary>
@@ -218,7 +218,7 @@ namespace OSharp.Data
         public static void FileExists(string filename, string paramName = null)
         {
             NotNull(filename, paramName);
-            Require<FileNotFoundException>(File.Exists(filename), string.Format(Resources.ParameterCheck_FileNotExists, filename));
+            Required<FileNotFoundException>(File.Exists(filename), string.Format(Resources.ParameterCheck_FileNotExists, filename));
         }
 
         /// <summary>
