@@ -9,11 +9,12 @@
 
 using Microsoft.Extensions.DependencyInjection;
 using OSharp.Audits;
-using OSharp.Core.Functions;
 using OSharp.Entity;
 using System;
 using System.Collections.Concurrent;
 using System.Security.Claims;
+
+using OSharp.Authorization.Functions;
 
 
 namespace OSharp.Dependency
@@ -22,7 +23,7 @@ namespace OSharp.Dependency
     /// 基于Scoped生命周期的数据字典，可用于在Scoped环境中传递数据
     /// </summary>
     [Dependency(ServiceLifetime.Scoped, AddSelf = true)]
-    public class ScopedDictionary : ConcurrentDictionary<string, object>, IDisposable
+    public sealed class ScopedDictionary : ConcurrentDictionary<string, object>, IDisposable
     {
         /// <summary>
         /// 获取或设置 当前执行的功能
